requirements.txt
requests==2.31.0
APScheduler==3.10.4
python-dateutil==2.8.2
pytz==2024.1
python-dotenv==1.0.0